Add expand symint python test and re-enable some previously disabled dynamic shape tests. (#4581)
* add expand_symint test to replace the cpp counterpart test.

* re-enable test AtenXlaTensorTest.TestNonzero

* re-enable test TestDynamicShape.test_nonzero_cast

* fix linter

* remove print

* fix pr comment

* add test_utils

* fix linter

* trying to fix more tests.

* fix linter

* fixed the error

* fix linter
