+++ test/cpp/test_aten_xla_tensor.cpp
@@ -5207,66 +5207,6 @@ static c10::SymInt make_symint(const torch::lazy::NodePtr& p) {
       static_cast<c10::SymNode>(c10::make_intrusive<XLASymNodeImpl>(p)));
 }
 
-TEST_F(AtenXlaTensorTest, TestExpandSymIntSymbolic) {
-  torch::Tensor a = torch::ones({3, 4}, torch::TensorOptions(torch::kFloat));
-  torch::Tensor b = a.expand({2, 3, 4}, /*implicit=*/false);
-
-  // create a symbolic symint with value 2 and upperbound 2. This symint is
-  // symbolic(since it wraps around a sizeNode) but not really dynamic.
-  torch::lazy::Value scalar_value =
-      torch::lazy::Value(ScalarOp(1.0, xla::F32), 0);
-  std::vector<int64_t> target_size = {2, 3, 4};
-  torch::lazy::NodePtr expand_node =
-      torch::lazy::MakeNode<Expand>(scalar_value, target_size);
-  torch::lazy::Value expand_value = torch::lazy::Value(expand_node, 0);
-  torch::lazy::NodePtr size_node =
-      torch::lazy::MakeNode<SizeNode>(expand_value, /*dim=*/0);
-  // This is not a dynamic size from xla perspective but it is a symint that
-  // wraps around a SizeNode instead of a scalar.
-  c10::SymInt dynamic_symint = make_symint(size_node);
-
-  ForEachDevice([&](const torch::Device& device) {
-    torch::Tensor xla_a = CopyToDevice(a, device);
-    torch::Tensor xla_b = xla_a.expand_symint(
-        c10::SymIntArrayRef({dynamic_symint, c10::SymInt(3), c10::SymInt(4)}),
-        /*implicit=*/false);
-    EXPECT_EQ(ToCpuTensor(xla_b).sum().item().toInt(), 24);
-    AllClose(b, xla_b);
-  });
-
-  ExpectCounterNotChanged("aten::.*", cpp_test::GetIgnoredCounters());
-  ExpectCounterChanged("xla::expand_symint", cpp_test::GetIgnoredCounters());
-}
-
-TEST_F(AtenXlaTensorTest, TestExpandSymIntDynamic) {
-  torch::Tensor a = torch::ones({3, 4}, torch::TensorOptions(torch::kFloat));
-  torch::Tensor b = a.expand({2, 3, 4}, /*implicit=*/false);
-
-  // use non_zero to create a symbolic symint with upper bound 100 and real
-  // value 2
-  int64_t num_non_zero_element = 2;
-  int64_t num_row = 10;
-  int64_t num_col = 10;
-  torch::lazy::NodePtr nonzero_node =
-      CreateNonZeroNode2d(num_non_zero_element, num_row, num_col);
-
-  torch::lazy::NodePtr size_node_nonzero_0 =
-      torch::lazy::MakeNode<SizeNode>(nonzero_node, 0);
-  c10::SymInt dynamic_symint = make_symint(size_node_nonzero_0);
-
-  ForEachDevice([&](const torch::Device& device) {
-    torch::Tensor xla_a = CopyToDevice(a, device);
-    torch::Tensor xla_b = xla_a.expand_symint(
-        c10::SymIntArrayRef({dynamic_symint, c10::SymInt(3), c10::SymInt(4)}),
-        /*implicit=*/false);
-    EXPECT_EQ(ToCpuTensor(xla_b).sum().item().toInt(), 24);
-    AllClose(b, xla_b);
-  });
-
-  ExpectCounterNotChanged("aten::.*", cpp_test::GetIgnoredCounters());
-  ExpectCounterChanged("xla::expand_symint", cpp_test::GetIgnoredCounters());
-}
-
 TEST_F(AtenXlaTensorTest, TestEye) {
   int n = 5;
   ForEachDevice([&](const torch::Device& device) {
@@ -5369,13 +5309,6 @@ TEST_F(AtenXlaTensorTest, TestOneIndexTransfer) {
   }
 }
 
-// Temporarily disable test. Original issue
-// https://github.com/pytorch/xla/issues/4501 has been resolved. The next error
-// is https://gist.github.com/vanbasten23/b3a79e0cc7f17edc0018eb83cdd5d738 (see
-// https://github.com/pytorch/xla/issues/4432 for more info). The next error
-// happens on TPU but not on CPU.
-//
-/*
 TEST_F(AtenXlaTensorTest, TestNonzero) {
   torch::Tensor a = torch::zeros({4, 2}, torch::TensorOptions(torch::kFloat));
   a[0][1] = 1.0;
@@ -5395,7 +5328,6 @@ TEST_F(AtenXlaTensorTest, TestNonzero) {
     ResetCounters();
   });
 }
-*/
 
 TEST_F(AtenXlaTensorTest, TestMaskedSelect) {
   torch::Tensor a = torch::rand({3, 5}, torch::TensorOptions(torch::kFloat));
